+++ frontend/sidedish/src/App.jsx
@@ -1,6 +1,6 @@
-import Main from "./components/Main";
+import Main from "./components/main/Main";
 import { Reset } from "styled-reset";
-import Header from "./components/Header";
+import Header from "./components/header/Header";
 
 const App = () => {
   return (
